api/setup/requirements.txt
pyjwt==1.7.1